// src/app/models/landRecord.ts
export class LandRecord {
    id?: string;
    division: string;
    district: string;
    taluk: string;
    village: string;
    name_of_scheme: string;
    land_alienated: string;
    date_of_enterupon_permission: string;
    sf_number: string;
    go_details: string;
    ownership_land_revenue: string;
    utilization_land: string;
    GOFile: any;
    otherFile: any;
}